Use Eloop and sockets for WPA CLI
In Linux WPA CLI is a standalone application that uses sockets to
communicate with ctrl_iface (uses Eloop), in Zephyr we have replaced
sockets with direct function calls, but this introduces concurrency as
we don't go through Eloop.

Implement Zephyr's version of WPA CLI and control interface using
socketpair UNIX style sockets, here WPA CLI can be invoked as a standard
function rather than an executable.

Also, as WPA CLI is meant for internal usage, remove exposing of limited
set of commands and instead make it a pass through to accept any
command.

Also, restore original wpa_cli as that is not used anymore.

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>